Store Downloads
Added option to auto download files or have them stored in either indexed db or a app downloads directory to be downloaded/moved to the correct location upon user request to actually download the file instead of automatically dropping in. Still need to build the menu that shows the current files in storage. Need to build a setting for when the files should be autodeleted from app storage if they have not been interacted with.